Move internal modules to separate namespace
